Collapse fuzz tests into test suites
